ci/requirements.txt
azure-identity==1.5.0
azure-mgmt-compute==12.0.0
azure-mgmt-network==10.2.0
azure-mgmt-resource==10.0.0
boto3==1.15.0
cachetools==4.1.1
pyyaml==5.3.1
slugid==2.0.0
taskcluster==38.0.6
urllib3==1.25.10
